MAINT: move all masked array matrix tests to matrixlib.
Further progress in isolating matrix in preparation of its
deprecation.  There is one place left with an explicit reference
to matrix (in MaskedArray.count), which is to be solved later.
